Avoid wasting the real liquibase version number for OSSRH deployment testing.
The plan is to make the feature version number follow the liquibase
release number.

Experience from previous OSSRH releases is that I always miss
something during release, so I will do a staging release, and perhaps
a full release, using the 1.0.0 (and 1.0.x as needed).

So temporarily downgrading the version number.